+++ libcnb-cargo/src/cross_compile.rs
@@ -9,16 +9,30 @@ use which::which;
 pub fn cross_compile_assistance(target_triple: impl AsRef<str>) -> CrossCompileAssistance {
     // Background: https://omarkhawaja.com/cross-compiling-rust-from-macos-to-linux/
     if target_triple.as_ref() == X86_64_UNKNOWN_LINUX_MUSL && cfg!(target_os = "macos") {
-        let gcc_binary_name = "x86_64-linux-musl-gcc";
+        // There is more than just one binary name here since we also support the binary name for
+        // an older version cross_compile_assistance which suggested installing a different
+        // toolchain.
+        let possible_gcc_binary_names =
+            vec!["x86_64-unknown-linux-musl-gcc", "x86_64-linux-musl-gcc"];
 
-        match which(gcc_binary_name) {
-            Ok(gcc_binary_path) => {
+        possible_gcc_binary_names
+            .iter()
+            .find_map(|binary_name| which(binary_name).ok())
+            .map_or_else(|| CrossCompileAssistance::HelpText(String::from(
+                r#"For cross-compilation from macOS to x86_64-unknown-linux-musl, a C compiler and
+linker for the target platform must be installed on your computer.
+
+The easiest way to install the required cross-compilation toolchain is to run:
+brew install messense/macos-cross-toolchains/x86_64-unknown-linux-musl
+
+For more information, see:
+https://github.com/messense/homebrew-macos-cross-toolchains"#,
+            )), |gcc_binary_path| {
                 CrossCompileAssistance::Configuration {
                     cargo_env: vec![
                         (
                             // Required until Cargo can auto-detect the musl-cross gcc/linker itself,
-                            // since otherwise it checks for a binary named 'musl-gcc'
-                            // not 'x86_64-linux-musl-gcc':
+                            // since otherwise it checks for a binary named 'musl-gcc':
                             // https://github.com/FiloSottile/homebrew-musl-cross/issues/16
                             // https://github.com/rust-lang/cargo/issues/4133
                             OsString::from("CARGO_TARGET_X86_64_UNKNOWN_LINUX_MUSL_LINKER"),
@@ -32,17 +46,7 @@ pub fn cross_compile_assistance(target_triple: impl AsRef<str>) -> CrossCompileA
                         ),
                     ],
                 }
-            }
-            Err(_) => CrossCompileAssistance::HelpText(String::from(
-                r#"For cross-compilation from macOS to x86_64-unknown-linux-musl, a C compiler and
-linker for the target platform must be installed on your computer.
-
-The easiest way to install 'x86_64-linux-musl-gcc' is to follow the instructions in the linked
-GitHub repository:
-
-https://github.com/FiloSottile/homebrew-musl-cross"#,
-            )),
-        }
+            })
     } else if target_triple.as_ref() == X86_64_UNKNOWN_LINUX_MUSL && cfg!(target_os = "linux") {
         match which("musl-gcc") {
             Ok(_) => CrossCompileAssistance::Configuration { cargo_env: vec![] },